Implement User-Based UI + Permissions

+++ src/components/pages/AppointmentsPage.tsx
@@ -1,11 +1,18 @@
 // Copyright 2022 Social Fabric, LLC
 
-import React, { useContext, useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
+import {
+  createPermission,
+  deletePermission,
+  updatePermission,
+} from '../../auth/permissions';
 import {
   Appointment,
   AppointmentsContext,
   emptyAppointment,
 } from '../../data/appointments';
+import { CounselorsContext, emptyCounselor } from '../../data/counselors';
+import { LoggedInUserContext } from '../../data/users';
 import CreateAppointmentModal from '../modals/CreateAppointmentModal';
 import EditAppointmentModal from '../modals/EditAppointmentModal';
 import Navbar from '../navbar/Navbar';
@@ -17,6 +24,8 @@ const AppointmentsPage: React.FC = () => {
     delete: deleteAppointment,
     update: updateAppointment,
   } = useContext(AppointmentsContext);
+  const { data: counselors } = useContext(CounselorsContext);
+  const { loggedInUser } = useContext(LoggedInUserContext);
 
   const [isCreateAppointmentModalOpen, setIsCreateAppointmentModalOpen] =
     useState<boolean>(false);
@@ -24,26 +33,70 @@ const AppointmentsPage: React.FC = () => {
     useState<boolean>(false);
   const [initialAppointment, setInitialAppointment] =
     useState<Appointment>(emptyAppointment);
+  const [isCreateAppointmentAllowed, setIsCreateAppointmentAllowed] =
+    useState<boolean>(false);
+  const [isDeleteAppointmentAllowed, setIsDeleteAppointmentAllowed] =
+    useState<boolean>(false);
+  const [isUpdateAppointmentAllowed, setIsUpdateAppointmentAllowed] =
+    useState<boolean>(false);
+
+  useEffect(() => {
+    setIsCreateAppointmentAllowed(
+      createPermission(loggedInUser.role, 'appointment')
+    );
+    setIsDeleteAppointmentAllowed(
+      deletePermission(loggedInUser.role, 'appointment')
+    );
+    setIsUpdateAppointmentAllowed(
+      updatePermission(loggedInUser.role, 'appointment')
+    );
+  }, [loggedInUser.role]);
 
   const onCreateAppointmentSubmit = (appointment: Appointment) => {
-    addAppointment(appointment);
+    if (isCreateAppointmentAllowed) {
+      addAppointment(appointment);
+    }
     setIsCreateAppointmentModalOpen(false);
   };
 
   const onEditAppointmentSubmit = (appointment: Appointment) => {
-    updateAppointment(appointment);
+    if (isUpdateAppointmentAllowed) {
+      updateAppointment(appointment);
+    }
     setIsEditAppointmentModalOpen(false);
   };
 
   const onAppointmentDeleteClicked = (appointmentToDelete: Appointment) => {
-    if (window.confirm('Delete this appointment?')) {
+    if (
+      isDeleteAppointmentAllowed &&
+      window.confirm('Delete this appointment?')
+    ) {
       deleteAppointment(appointmentToDelete);
     }
   };
 
   const onAppointmentEditClicked = (appointmentToEdit: Appointment) => {
-    setInitialAppointment(appointmentToEdit);
-    setIsEditAppointmentModalOpen(!isEditAppointmentModalOpen);
+    if (isUpdateAppointmentAllowed) {
+      setInitialAppointment(appointmentToEdit);
+      setIsEditAppointmentModalOpen(!isEditAppointmentModalOpen);
+    }
+  };
+
+  const onAppointmentEmailClicked = (appointmentToEmail: Appointment) => {
+    const appointmentCounselor =
+      counselors.find(
+        counselor =>
+          counselor.counselorRef.id === appointmentToEmail.counselorId
+      ) || emptyCounselor;
+    let mailToUrl = `mailto:${appointmentCounselor.email}`;
+
+    for (const participant of appointmentToEmail.participants) {
+      mailToUrl += `,${participant.email}`;
+    }
+
+    mailToUrl += `?subject=${encodeURIComponent(appointmentToEmail.title)}`;
+
+    window.open(mailToUrl);
   };
 
   return (
@@ -53,28 +106,35 @@ const AppointmentsPage: React.FC = () => {
       </nav>
       <h1>Appointments</h1>
       <>
-        <button
-          type="button"
-          onClick={() => setIsCreateAppointmentModalOpen(true)}
-        >
-          Add Appointment
-        </button>
+        {isCreateAppointmentAllowed && (
+          <button
+            type="button"
+            onClick={() => setIsCreateAppointmentModalOpen(true)}
+          >
+            Add Appointment
+          </button>
+        )}
         <AppointmentsTable
           onDeleteClicked={onAppointmentDeleteClicked}
           onEditClicked={onAppointmentEditClicked}
+          onEmailClicked={onAppointmentEmailClicked}
         />
-        <CreateAppointmentModal
-          isOpen={isCreateAppointmentModalOpen}
-          onClose={() => setIsCreateAppointmentModalOpen(false)}
-          onAppointmentAdded={onCreateAppointmentSubmit}
-          initialAppointment={emptyAppointment}
-        />
-        <EditAppointmentModal
-          isOpen={isEditAppointmentModalOpen}
-          onClose={() => setIsEditAppointmentModalOpen(false)}
-          onAppointmentAdded={onEditAppointmentSubmit}
-          initialAppointment={initialAppointment}
-        />
+        {isCreateAppointmentAllowed && (
+          <CreateAppointmentModal
+            isOpen={isCreateAppointmentModalOpen}
+            onClose={() => setIsCreateAppointmentModalOpen(false)}
+            onAppointmentAdded={onCreateAppointmentSubmit}
+            initialAppointment={emptyAppointment}
+          />
+        )}
+        {isUpdateAppointmentAllowed && (
+          <EditAppointmentModal
+            isOpen={isEditAppointmentModalOpen}
+            onClose={() => setIsEditAppointmentModalOpen(false)}
+            onAppointmentAdded={onEditAppointmentSubmit}
+            initialAppointment={initialAppointment}
+          />
+        )}
       </>
     </div>
   );
