+++ src/components/pages/UsersPage.tsx
@@ -1,7 +1,17 @@
 // Copyright 2022 Social Fabric, LLC
 
-import React, { useContext, useState } from 'react';
-import { emptyUser, User, UsersContext } from '../../data/users';
+import React, { useContext, useEffect, useState } from 'react';
+import {
+  createPermission,
+  deletePermission,
+  updatePermission,
+} from '../../auth/permissions';
+import {
+  emptyUser,
+  LoggedInUserContext,
+  User,
+  UsersContext,
+} from '../../data/users';
 import CreateUserModal from '../modals/CreateUserModal';
 import EditUserModal from '../modals/EditUserModal';
 import Navbar from '../navbar/Navbar';
@@ -13,33 +23,53 @@ const UsersPage = () => {
     delete: deleteUser,
     update: updateUser,
   } = useContext(UsersContext);
+  const { loggedInUser } = useContext(LoggedInUserContext);
+
   const [isCreateUserModalOpen, setIsCreateUserModalOpen] =
     useState<boolean>(false);
   const [isEditUserModalOpen, setIsEditUserModalOpen] =
     useState<boolean>(false);
   const [modalUser, setModalUser] = useState<User>(emptyUser);
+  const [isCreateUserAllowed, setIsCreateUserAllowed] =
+    useState<boolean>(false);
+  const [isDeleteUserAllowed, setIsDeleteUserAllowed] =
+    useState<boolean>(false);
+  const [isUpdateUserAllowed, setIsUpdateUserAllowed] =
+    useState<boolean>(false);
+
+  useEffect(() => {
+    setIsCreateUserAllowed(createPermission(loggedInUser.role, 'user'));
+    setIsDeleteUserAllowed(deletePermission(loggedInUser.role, 'user'));
+    setIsUpdateUserAllowed(updatePermission(loggedInUser.role, 'user'));
+  }, [loggedInUser.role]);
 
   const handleUserAdded = (user: User) => {
-    addUser(user);
+    if (isCreateUserAllowed) {
+      addUser(user);
+    }
   };
 
   const handleUserEdited = (user: User) => {
-    updateUser(user);
+    if (isUpdateUserAllowed) {
+      updateUser(user);
+    }
   };
 
   const onUserDeleteClicked = (userToDelete: User) => {
-    if (window.confirm('Delete this user?')) {
+    if (isDeleteUserAllowed && window.confirm('Delete this user?')) {
       deleteUser(userToDelete);
     }
   };
 
   const onUserEditClicked = (userToEdit: User) => {
-    setModalUser(userToEdit);
-    setIsEditUserModalOpen(true);
+    if (isUpdateUserAllowed) {
+      setModalUser(userToEdit);
+      setIsEditUserModalOpen(true);
+    }
   };
 
-  const onUserEmailClicked = (userToEdit: User) => {
-    window.location.href = `mailto:${userToEdit.email}`;
+  const onUserEmailClicked = (userToEmail: User) => {
+    window.open(`mailto:${userToEmail.email}`);
   };
 
   return (
@@ -49,20 +79,29 @@ const UsersPage = () => {
       </nav>
       <h1>Users</h1>
       <>
-        <button type="button" onClick={() => setIsCreateUserModalOpen(true)}>
-          Add User
-        </button>
-        <CreateUserModal
-          isOpen={isCreateUserModalOpen}
-          onUserAdded={handleUserAdded}
-          onClose={() => setIsCreateUserModalOpen(false)}
-        />
-        <EditUserModal
-          isOpen={isEditUserModalOpen}
-          onUserEdited={handleUserEdited}
-          onClose={() => setIsEditUserModalOpen(false)}
-          initialUser={modalUser}
-        />
+        {isCreateUserAllowed && (
+          <>
+            <button
+              type="button"
+              onClick={() => setIsCreateUserModalOpen(true)}
+            >
+              Add User
+            </button>
+            <CreateUserModal
+              isOpen={isCreateUserModalOpen}
+              onUserAdded={handleUserAdded}
+              onClose={() => setIsCreateUserModalOpen(false)}
+            />
+          </>
+        )}
+        {isUpdateUserAllowed && (
+          <EditUserModal
+            isOpen={isEditUserModalOpen}
+            onUserEdited={handleUserEdited}
+            onClose={() => setIsEditUserModalOpen(false)}
+            initialUser={modalUser}
+          />
+        )}
         <UsersTable
           onDeleteClicked={onUserDeleteClicked}
           onEditClicked={onUserEditClicked}
